Explaining Associations
A rule takes the form X ! Y with X; Y  I and X \ Y = ;. X and Y are called itemsets.

The support supp(X) of an itemset X is defined as the proportion of transactions in the data set which contain the itemset. In the example database in Figure 1, the itemset {milk, bread} has a support of 2/5 = 0.4 since it occurs in 40% of all transactions (2 out of 5 transactions).

The confidence of a rule is defined conf(X ) Y ) = supp(X [ Y )/supp(X). For example, the rule {milk, bread} ) {butter} has a confidence of 0.2/0.4 = 0.5 in the database in Figure 1, which means that for 50% of the transactions containing milk and bread the rule is correct. confidence can be interpreted as an estimate of the probability P(Y |X), the probability of
finding the RHS of the rule in transactions under the condition that these transactions also contain the LHS.  It’s the share of transactions containing Y in all the transactions containing X.

The lift of a rule is defined as lift(X ) Y ) = supp(X [ Y )/(supp(X)supp(Y )), and can be interpreted as the deviation of the support of the whole rule from the support expected under independence given the supports of the LHS and the RHS. Greater lift values indicate stronger associations.